requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-heroku==0.3.1
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
pycryptodomex==3.9.8
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
static3==0.7.0
stream-python==4.0.0
urllib3==1.25.11
whitenoise==5.2.0
